requirements_dev.txt
django_admin_generator
